Refs codership/QA#90 Enable 32-bit MySQL-wsrep builds

+++ storage/innobase/lock/lock0lock.cc
@@ -2323,8 +2323,8 @@ lock_rec_add_to_queue(
 			if (wsrep_thd_is_BF(trx->mysql_thd, FALSE)) {
 				if (wsrep_debug) {
 					fprintf(stderr, 
-						"BF skipping wait: %lu\n", 
-						trx->id);
+						"BF skipping wait: %llu\n",
+						(unsigned long long) trx->id);
 					lock_rec_print(stderr, lock);
 				}
 		  } else
